Refactor sample code
