+++ webapp/src/components/Navbar.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
 
-import './Navbar.css';
-
 function Navbar () {
 
     return (
