+++ webapp/src/pages/tests/HeaderTest.js
@@ -1,4 +1,4 @@
-import Header from "../../components/header/Header";
+import Header from "../../components/Header";
 
 function HeaderTest(){
