+++ webapp/src/pages/tests/NavbarTest.js
@@ -1,4 +1,4 @@
-import Navbar from "../../components/navbar/Navbar";
+import Navbar from "../../components/Navbar";
 
 function NavbarTest(){
